fix: add registration submit btn, fix the code based on comments
