+++ apps/client/src/pages/RegistrationPage/RegistrationPage.ts
@@ -8,7 +8,7 @@ export default class RegistrationPage extends ContainerComponent {
     super({
       tag: 'section',
       classes:
-        'relative w-screen h-screen px-10 py-5 flex flex-col justify-between items-center bg-[url("src/assets/images/bg-1.png")] bg-center bg-cover bg-no-repeat',
+        'relative w-screen h-screen px-10 py-5 flex flex-col justify-center items-center bg-[url("src/assets/images/bg-1.png")] bg-center bg-cover bg-no-repeat',
     });
 
     this.render();
